chore: add the mongo DB instance to the main application
