+++ .gitignore
@@ -1,1 +1,15 @@
 app.env
+*.exe
+*.exe~
+*.dll
+*.so
+*.dylib
+
+*.test
+
+*.out
+
+go.work
+go.work.sum
+
+.env
