+++ Makefile
@@ -1,8 +1,29 @@
 GO=go
 GOTEST=${GO} test -v
-COLORIZE ?= | $(SED) 's/PASS/✅ PASS/g' | $(SED) 's/FAIL/❌ FAIL/g' | $(SED) 's/SKIP/🔕 SKIP/g'
-
+COLORIZE ?= | sed 's/PASS/✅ PASS/g' | sed 's/FAIL/❌ FAIL/g' | sed 's/SKIP/🔕 SKIP/g'
 
 .PHONY: test
 test:
 	bash -c "set -e; set -o pipefail; $(GOTEST) . $(COLORIZE)"
+
+# Start Docker Compose
+.PHONY: docker-up
+docker-up:
+	docker compose up -d
+
+# Stop and remove containers
+.PHONY: docker-down
+docker-down:
+	docker compose down
+
+# Build and run MongoDB manually (without Compose)
+.PHONY: docker-container
+docker-container:
+	docker build -t my-mongodb .
+	docker run -d -p 27017:27017 --name mongodb my-mongodb
+
+
+.PHONY: clean-container
+clean-container:
+	docker stop mongodb
+	docker rm mongodb
