+++ docker-compose.yml
@@ -1,0 +1,17 @@
+# this should only be used in development mode
+version: "3.8"
+services:
+  mongodb:
+    image: mongo:7.0
+    container_name: mongodb_container
+    restart: always
+    environment:
+      MONGO_INITDB_ROOT_USERNAME: root
+      MONGO_INITDB_ROOT_PASSWORD: password
+    ports:
+      - "27017:27017"
+    volumes:
+      - mongodb_data:/data/db
+
+volumes:
+  mongodb_data:
