+++ internal/repository/base.go
@@ -1,0 +1,16 @@
+package repository
+
+import (
+	dbinterface "github.com/telexintegrations/support-ai/internal/repository/dbInterface"
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+type DatabaseManager struct {
+	Mongo dbinterface.MongoManager
+}
+
+var DB = DatabaseManager{}
+
+func ConnectDB(client *mongo.Client) *DatabaseManager {
+	return &DatabaseManager{}
+}
