+++ internal/repository/dbInterface/db_interface.go
@@ -1,0 +1,9 @@
+package dbinterface
+
+import (
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+type MongoManager interface {
+	DB() *mongo.Client
+}
