+++ internal/repository/mongo/manager.go
@@ -1,21 +1,17 @@
 package mongo
 
-import "go.mongodb.org/mongo-driver/mongo"
+import (
+	"go.mongodb.org/mongo-driver/mongo"
+)
 
 type MongoDB struct {
-	mongoClient *mongo.Client
+	MongoClient *mongo.Client
 }
 
 type MongoManager interface {
 	DB() *mongo.Client
 }
 
-func NewMongoManager(client *mongo.Client) MongoManager {
-	return &MongoDB{
-		mongoClient: client,
-	}
-}
-
 func (m *MongoDB) DB() *mongo.Client {
-	return m.mongoClient
+	return m.MongoClient
 }
